Refactor the debug bundle generator to be ready to use from engine
